requirements.txt
Flask
SQLAlchemy
pandas
openpyxl
XlsxWriter
gunicorn
psycopg2-binary
python-dotenv
werkzeug